sync code from release-3.0.0-rc4

+++ bcos-executor/src/precompiled/solidity/BfsPrecompiled.sol
@@ -9,9 +9,9 @@ pragma experimental ABIEncoderV2;
     }
 
 contract BfsPrecompiled {
-    function list(string memory absolutePath) public view returns (int256,BfsInfo[] memory){}
-    function mkdir(string memory absolutePath) public returns (int256){}
-    function link(string memory name, string memory version, address _address, string memory _abi) public returns (int256){}
+    function list(string memory absolutePath) public view returns (int32,BfsInfo[] memory){}
+    function mkdir(string memory absolutePath) public returns (int32){}
+    function link(string memory name, string memory version, address _address, string memory _abi) public returns (int32){}
     function readlink(string memory absolutePath) public view returns (address) {}
-    function touch(string memory absolutePath, string memory fileType) public returns (int256){}
+    function touch(string memory absolutePath, string memory fileType) public returns (int32){}
 }
